+++ MoltenVK/MoltenVK/Utility/MVKSmallVector.h
@@ -43,11 +43,11 @@ using MVKSmallVector = std::vector<T>;
 // If you don't need any inline storage use:
 //  MVKSmallVector<int> v;   // this is essentially the same as using std::vector
 //
-// The per-instance memory overhead of MVKSmallVector (16 bytes) is smaller than MVKVector (40 bytes)
-// and std::vector (24 bytes), but MVKSmallVector lacks the polymorphism of MVKVector (or std::vector),
-// that allows them to be passed around to functions without reference to the pre-allocation size.
-// MVKSmallVector supports the contents() function to derive an MVKArrayRef from its contents,
-// which can be passed around without reference to the pre-allocaton size.
+// The per-instance memory overhead of MVKSmallVector (16 bytes) is smaller than std::vector (24 bytes),
+// and MVKSmallVector lacks the polymorphism of std::vector, that allows them to be passed around to
+// functions without reference to the pre-allocation size. MVKSmallVector supports the contents()
+// function to derive an MVKArrayRef from its contents, which can be passed around without reference
+// to the pre-allocaton size.
 
 #include "MVKSmallVectorAllocator.h"
 #include "MVKFoundation.h"